테스트 종료 : 개발 브랜치로 병합

+++ src/main/java/com/fortest/orderdelivery/app/domain/review/dto/ReviewGetResponseDto.java
@@ -21,7 +21,7 @@ public class ReviewGetResponseDto {
 
     private String contents;
 
-    private Long rate;
+    private Integer rate;
 
     @JsonProperty("created-at")
     private String createdAt;
